Update SS14.GithubApiHelper dependency Update octokit Fix github api related IDs being int instead of long Change usages of .Find to .Where Fix dbset nullability

+++ SS14.MapServer/Services/Github/GithubApiService.cs
@@ -1,4 +1,5 @@
 ﻿using Octokit;
+using SS14.GithubApiHelper.Exceptions;
 using SS14.GithubApiHelper.Services;
 using SS14.MapServer.Configuration;
 using SS14.MapServer.Exceptions;
@@ -32,7 +33,7 @@ public sealed class GithubApiService : AbstractGithubApiService
         return compareResult.Files.Select(file => file.Filename);
     }
 
-    public async Task<int?> CreateCommentWithTemplate(InstallationIdentifier installation, IssueIdentifier issue, string templateName, object model)
+    public async Task<long?> CreateCommentWithTemplate(InstallationIdentifier installation, IssueIdentifier issue, string templateName, object model)
     {
         if (!await CheckRateLimit(installation))
             return null;
@@ -51,7 +52,7 @@ public sealed class GithubApiService : AbstractGithubApiService
         return null;
     }
 
-    public async Task UpdateCommentWithTemplate(InstallationIdentifier installation, IssueIdentifier issue, int commentId, string templateName, object model)
+    public async Task UpdateCommentWithTemplate(InstallationIdentifier installation, IssueIdentifier issue, long commentId, string templateName, object model)
     {
         if (!await CheckRateLimit(installation))
             return;
